# app/Http/Requests/Compound/CompoundStoreRequest.php
<?php

namespace App\Http\Requests\Compound;

use Illuminate\Foundation\Http\FormRequest;

class CompoundStoreRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array
     */
    public function rules(): array
    {
        return [
            'name' => 'required|max:255',
        ];
    }
}