# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación Atlas",
        "description": "Documentación WPARK - API",
        "contact": {
            "email": "lchirinos@acercapartners.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/compounds": {
            "get": {
                "tags": [
                    "Compounds"
                ],
                "summary": "Compounds List",
                "description": "List of compounds",
                "operationId": "indexCompounds",
                "responses": {
                    "200": {
                        "description": "Compound list Successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Compounds"
                ],
                "summary": "Create New Compound",
                "description": "Create New Compound",
                "operationId": "storeCompound",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompoundStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create New Compound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/compounds/{id}": {
            "get": {
                "tags": [
                    "Compounds"
                ],
                "summary": "Show Compound Details",
                "description": "Show Compound Details",
                "operationId": "showCompound",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show Compound Details"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Compounds"
                ],
                "summary": "Update Compound",
                "description": "Update Compound",
                "operationId": "updateCompound",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompoundUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Compound"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compounds"
                ],
                "summary": "Delete Compound",
                "description": "Delete Compound",
                "operationId": "destroyCompound",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete Compound successfully"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Compounds"
                ],
                "summary": "Restore Compound",
                "description": "Restore Compound",
                "operationId": "restoreCompound",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Compound restored successfully"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/parking-design": {
            "post": {
                "tags": [
                    "ParkingDesigns"
                ],
                "summary": "Create New Parking",
                "description": "Create New Parking and all associated rows and slots",
                "operationId": "storeParkingDesign",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParkingDesignStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create New ParkingDesign"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompoundStoreRequest": {
                "title": "Compound Store Request",
                "description": "Compound Store request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre de la campa",
                        "type": "string",
                        "maxLength": 255,
                        "example": "FORD VALENCIA"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CompoundStoreRequest"
                }
            },
            "CompoundUpdateRequest": {
                "title": "Compound Update Request",
                "description": "Compound Update request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre de la campa",
                        "type": "string",
                        "maxLength": 255,
                        "example": "FORD VALENCIA"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "CompoundUpdateRequest"
                }
            },
...
            "ParkingDesignStoreRequest": {
                "title": "Parking Design Store Request",
                "description": "Parking Design Store request body data",
                "required": [
                    "name",
                    "area_id",
                    "parking_type_id",
                    "start_row",
                    "end_row"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del parking",
                        "type": "string",
                        "maxLength": 255,
                        "example": "PU"
                    },
                    "area_id": {
                        "description": "Indica el área al que pertenece el parking",
                        "type": "integer",
                        "maxLength": 20,
                        "example": "1"
                    },
                    "parking_type_id": {
                        "description": "Indica el tipo de parking",
                        "type": "integer",
                        "maxLength": 20,
                        "example": "1"
                    },
                    "start_row": {
                        "description": "La fila del área en la que empieza el parking",
                        "type": "integer",
                        "maxLength": 10,
                        "example": "12"
                    },
                    "end_row": {
                        "description": "La fila del área en la que termina el parking",
                        "type": "integer",
                        "maxLength": 10,
                        "example": "22"
                    },
                    "comments": {
                        "description": "Comentarios sobre el parking",
                        "type": "string",
                        "example": "Este parking deberá cambiar de ubicación"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ParkingDesignStoreRequest"
                }
            },
...
        },
        "responses": {
...
            "Forbidden": {
                "description": "Permission Error."
            },
            "InternalServerError": {
                "description": "Internal Server Error"
            },
            "NotFound": {
                "description": "Resource Not Found."
            },
            "Unauthorized": {
                "description": "Unauthorized Error."
            },
            "UnprocessableEntity": {
                "description": "Validation Error."
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "id, eg; 1",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}